Implement exponential back-off & match AWS SDK backoff params
