+++ typescript/src/backends/dynamo.ts
@@ -39,7 +39,7 @@ export class Dynamo implements Store {
   }
 
   public async get(key: MemKey): Promise<Uint8Array | undefined> {
-    return await this.getWithRetry(key, 5, 100, 2);
+    return await this.getWithRetry(key, 30, 25, 2, 20000);
   }
 
   public async set(key: MemKey, value: Uint8Array): Promise<void> {
@@ -181,9 +181,10 @@ export class Dynamo implements Store {
 
   private async getWithRetry(
     key: MemKey,
-    maxRetries = 5,
-    waitTimeMs = 100,
-    factor = 2,
+    maxRetries: number,
+    baseDelay: number,
+    factor: number,
+    maxBackoffMs: number,
   ): Promise<Uint8Array | undefined> {
     for (let attempt = 0; attempt <= maxRetries; attempt++) {
       const { Item } = await this.client.send(
@@ -192,8 +193,9 @@ export class Dynamo implements Store {
       if (Item?.value) return Item.value;
 
       if (attempt < maxRetries) {
-        await new Promise((r) => setTimeout(r, waitTimeMs));
-        waitTimeMs *= factor;
+        await new Promise((r) =>
+          setTimeout(r, Math.min(baseDelay * factor ** attempt, maxBackoffMs)),
+        );
       }
     }
   }
